Ajout des cas 1, 2, 3, 4 et 5

+++ sources/ft_tab.c
@@ -6,7 +6,7 @@
 /*   By: acoezard <acoezard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 20:04:09 by axelcoezard       #+#    #+#             */
-/*   Updated: 2021/11/13 20:41:25 by acoezard         ###   ########.fr       */
+/*   Updated: 2021/11/13 23:41:03 by acoezard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,9 +30,9 @@ t_tab	convert_to_tab(t_list *stack)
 	return (tab);
 }
 
-size_t	find_index_in_tab(t_tab *tab, int value)
+int	find_index_in_tab(t_tab *tab, int value)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < tab->size)
@@ -46,9 +46,9 @@ size_t	find_index_in_tab(t_tab *tab, int value)
 
 void	sort_tab(t_tab *tab)
 {
-	size_t	cursor;
-	size_t	i;
-	size_t	min;
+	int		cursor;
+	int		i;
+	int		min;
 	int		tmp;
 
 	cursor = 0;
